rename openapi mod to parse mod
